pyproject.toml
[project]
name = "eldtrip-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.15",
    "django-cors-headers>=4.3",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "gunicorn>=22.0",
    "whitenoise>=6.7",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-django>=4.8",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
pythonpath = ["."]
